Add installation script
Usage: install <config set>

Where <config set> is one of the repo's subdirectory names (e.g.,
        `bash`, `py`).

Copies all the config files in the given config set to the home
directory, ready for use after a restart of the shell or a sourcing of
the files.

**WARNING**: The script will overwrite any config files in the home
directory whose name is `.` followed by any of the names of the source
files in the given config set (e.g., `.bashrc`, `.pyrc`). The script
does **NOT** warn or ask for confirmation before doing this. Be very
sure you really want to install **ALL** the files in the given set
before using the script. If not, install them manually.